Export the test into the folder
